test volume

+++ lib/ui/src/ui.h
@@ -293,7 +293,7 @@ LV_IMG_DECLARE( ui_img_iconcard_png);   // assets\iconCard.png
 
 extern bool addFinger;
 
-extern uint16_t volumeTime;
+extern unsigned long volumeTime;
 void ui_init(void);
 
 #ifdef __cplusplus
